requirements.txt
flask
python-dotenv
psycopg2-binary
pandas
alpha_vantage